8-11-22
Kind of a bust day. I got the tableview to "work" but it didnt really change the look like it does in their example. Gonna dig more into it. I created a test screen for the table view so I dont mess with our main work while I figure it out.